api-python/requirements.txt
PyPDF2==3.0.1
pdfminer.six==20221105
openai==1.55.3
python-dotenv==1.0.0
httpx==0.27.2
boto3==1.28.38